+++ package.json
@@ -1,26 +1,26 @@
 {
-  "name": "astro-i18n-starter",
-  "description": "A starter for Astro with i18n support",
-  "version": "0.0.1",
-  "private": false,
-  "license": "MIT",
-  "homepage": "https://astro-i18n-starter.pages.dev/",
-  "engines": {
-    "node": ">=18"
-  },
-  "scripts": {
-    "dev": "astro dev",
-    "start": "astro dev",
-    "build": "astro check && astro build",
-    "preview": "astro preview",
-    "astro": "astro"
-  },
-  "dependencies": {
-    "@astrojs/check": "^0.7.0",
-    "@astrojs/mdx": "^3.1.0",
-    "@astrojs/rss": "^4.0.6",
-    "@astrojs/sitemap": "^3.1.5",
-    "astro": "^4.10.2",
-    "typescript": "^5.4.5"
-  }
+	"name": "astro-i18n-starter",
+	"description": "A starter for Astro with i18n support",
+	"version": "2.0.0",
+	"private": false,
+	"license": "MIT",
+	"homepage": "https://astro-i18n-starter.pages.dev/",
+	"engines": {
+		"node": ">=20"
+	},
+	"scripts": {
+		"dev": "astro dev",
+		"start": "astro dev --open",
+		"build": "astro check && astro build",
+		"preview": "astro preview",
+		"astro": "astro"
+	},
+	"dependencies": {
+		"@astrojs/check": "^0.9.4",
+		"@astrojs/mdx": "^4.0.3",
+		"@astrojs/rss": "^4.0.11",
+		"@astrojs/sitemap": "^3.2.1",
+		"astro": "^5.1.3",
+		"typescript": "^5.4.5"
+	}
 }
