+++ src/blog/zh-cn/first-post/cover.png
